one last refinement (from looking at MetricsMvcEndpoint)
