+++ common/src/main/java/org/opentestsystem/rdw/ingest/status/StatusConfiguration.java
@@ -17,13 +17,20 @@ import java.util.List;
  * TODO - think about ProvidersIndicator; should we add additional framework so that we can provide a default
  * TODO   implementation that checks config-server, etc. while allowing extensibility?
  * </p>
+ * <p>
+ * NOTE: currently this just exposes the MVC bean. When it makes sense to directly expose StatusEndpoint
+ * (e.g. if there is need for a JMX wrapper) do so.
+ * Look at EndpointAutoConfiguration vs. EndpointWebMvcManagementContextConfiguration for a pattern to follow.
+ * </p>
+ *
+ * @see StatusEndpoint
  */
 @Configuration
 public class StatusConfiguration {
 
     @Bean
     public StatusMvcEndpoint statusMvcEndpoint(final List<StatusIndicator> statusIndicators) {
-        return new StatusMvcEndpoint(statusIndicators);
+        return new StatusMvcEndpoint(new StatusEndpoint(statusIndicators));
     }
 
     @Bean
