+++ common/src/main/java/org/opentestsystem/rdw/ingest/status/StatusEndpoint.java
@@ -1,0 +1,49 @@
+package org.opentestsystem.rdw.ingest.status;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.actuate.endpoint.AbstractEndpoint;
+import org.springframework.boot.actuate.endpoint.Endpoint;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.List;
+
+/**
+ * An {@link Endpoint} to expose web diagnostics.
+ *
+ * @see StatusConfiguration
+ */
+@ConfigurationProperties(prefix = "endpoints.status")
+public class StatusEndpoint extends AbstractEndpoint<SystemStatus> {
+
+    @Value("${spring.application.name}")
+    private String applicationName;
+
+    private List<StatusIndicator> statusIndicators;
+
+    public StatusEndpoint(final List<StatusIndicator> statusIndicators) {
+        super("status", true);
+        this.statusIndicators = statusIndicators;
+    }
+
+    @Override
+    public SystemStatus invoke() {
+        return invoke(0);
+    }
+
+    public SystemStatus invoke(final int level) {
+        return generateStatus(level);
+    }
+
+    private SystemStatus generateStatus(final int level) {
+        // TODO - set overall Rating based on indicator ratings
+
+        final SystemStatus.Builder builder = SystemStatus.builder()
+                .level(level)
+                .unit(applicationName);
+        for (final StatusIndicator statusIndicator : statusIndicators) {
+            if (level < statusIndicator.level()) continue;
+            builder.sub(statusIndicator.name(), statusIndicator.status());
+        }
+        return builder.build();
+    }
+}
